Extract Prompt Studio highlight helpers
